Saves an output file from measure2.sh to an HDF file
Uses a dictionary keyed by a tuple.  One of the regexes has gone away,
the one that words from the echo commands in measure2.  I should get
rid of the other regex as well, since I can look for the strings "(" and
" MB/s)" and grab the substring.
Now imports pandas and creates a MultiIndex data frame.
Stores the results in an HDF5 file.