Fixed REs to make more general
The reported bandwidth from the wget command is always scaled so that it
is 3 significant figures.  That means that (126 MB/s), (11.2 MB/s), and
(346 KB/s) are all legitimate answers.  So my regex has to be more
general and I have to know if I am dealing with MB/s or KB/s.  I also
have to multiply the value by either 1000 or 1000000.