4 working X-Y plots and IPv4 3D scatter plot works
The IPv6 3D scatter plot doesn't work, but that might be a data
issue.
The string "Bandwidth as a function of packet loss rate (%) and delay (msec) for IPv4"
is continued in such a way that there is white space between the words "packet" and "loss"
There should be unit strings on the axis for the 3D scatter plots.