Added a work around for the bug in tc
Reduced the number of test cases, and got rid of the 100% loss
test case until software development is complete.
Generates a time file which is much simpler, for use with the
time2MultIindex.py program, which generates the HDF5 format h5
file.

A "quick look" reveals that the data looks good.