global TCS retries, better elapsed time calculation
I wrote the function global_tcp_retries() and then I never called it!
So now I do, and the result is appended to the end of the line that
report generates.

I'm also getting insanely short elapsed time values, and I think the
problem is the s.send(data) doesn't actually block until it gets to
very large values of sizes of data.  Think about it for a moment: the
default window size is 64Kbytes.  So I have to send at least 64K before
I will even begin to fill the sliding window.  Which is why my data
analysis has seemed so random: it is.  I also moved where the start_time
and end_time are calculated.

s.close() was in the wrong place, so I moved it.